Moved the types file into the state/reducers folder, also had to rewrite paths on a few imports
